+++ src/FreeElephants/AltEra/Exception/AltEraExceptionInterface.php
@@ -5,6 +5,6 @@ namespace FreeElephants\AltEra\Exception;
 /**
  * @author samizdam
  */
-class MissedDependencyException extends RuntimeException
+interface AltEraExceptionInterface
 {
 }