+++ src/FreeElephants/AltEra/Exception/RuntimeException.php
@@ -5,6 +5,6 @@ namespace FreeElephants\AltEra\Exception;
 /**
  * @author samizdam
  */
-class RuntimeException extends \RuntimeException implements CalendarExceptionInterface
+class RuntimeException extends \RuntimeException implements AltEraExceptionInterface
 {
 }